Update person and rental to manage creation and show list
